Refactor main UI logic into MainView class
Moved all console UI and menu logic from main.py into a new utils/main_view.py MainView class, improving separation of concerns. main.py now only wires dependencies and delegates to MainView. Updated Game and GameFacade to clarify method names and docstrings, and to support the new UI structure.
